Users can view messages through a web interface
* Currently messages can only be seeded, or created through the console.
However this allows an interface to retrieve them.
* A new resource `message_view` is created. One cannot retrieve message
bodies via a GET request for security reasons, otherwise `messages`
would be the only `resource` for this application. The security concerns
include GET request addresses being logged, so including a
password/encryption key in the query string params would be a poor
choice.
* Including the form on a messages index page was avoided since rails 5
has separate auth tokens for each form by default (or possibly for each
controller). This can be disabled in the config, but it seems a good
idea.
http://blog.bigbinary.com/2016/01/11/per-form-csrf-token-in-rails-5.html

Other:
* Introducing capybara-webkit for integration testing
* Filtering sensitive information from logs (password, message body)
* Including launchy gem since it is useful for debugging feature specs
(type save_and_open_page to open current test html page in a browser
tab)
* CI already setup so this line is removed from the README
* Factory girl was setup to make testing easy
* jquery-rails gem was already removed so one cant require jquery.
Therefor it is removed from the js manifest file `application.js`